fix passport typescript namespace
